pyproject.toml
[project]
name = "autospook"
version = "0.2.0"
description = "OSINT research tool using LangChain and FastAPI"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langgraph>=0.0.40",
    "langsmith>=0.0.60",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.8",
    "langchain-community>=0.0.10",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "duckduckgo-search>=6.0.0",
    "wikipedia>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pip-audit>=2.6.0",
] 